+++ dom-tags/class-heading-tag.php
@@ -19,6 +19,15 @@ class HeadingTag extends \DomTags implements DomTagInterface {
 	private const TAG_TYPES = array('h1', 'h2', 'h3', 'h4', 'h5', 'h6');
 	
 	/**
+	 * The current tag type.
+	 * @since 1.2.0
+	 *
+	 * @access private
+	 * @var string
+	 */
+	private static $tag_type;
+	
+	/**
 	 * Construct the DOMtag.
 	 * @since 1.0.1
 	 *
@@ -27,12 +36,12 @@ class HeadingTag extends \DomTags implements DomTagInterface {
 	 * @return string
 	 */
 	public static function tag(?array $args = null): string {
-		$type = self::TAG_TYPES[1];
+		self::$tag_type = self::TAG_TYPES[1];
 		
-		if(isset($args['type']) && in_array($args['type'], self::TAG_TYPES, true))
-			$type = $args['type'];
+		if(isset($args['tag_type']) && in_array($args['tag_type'], self::TAG_TYPES, true))
+			self::$tag_type = $args['tag_type'];
 		
-		return parent::constructTag($type, self::props(), $args);
+		return parent::constructTag(self::$tag_type, self::props(), $args);
 	}
 	
 	/**