+++ dom-tags/class-iframe-tag.php
@@ -1,29 +1,29 @@
 <?php
 /**
- * The <fieldset> DOMtag.
- * @since 1.0.0
+ * The <iframe> DOMtag.
+ * @since 1.2.0
  *
  * @author Jace Fincham
  * @package DomTags
  */
 namespace DomTags;
 
-class FieldsetTag extends \DomTags implements DomTagInterface {
+class IframeTag extends \DomTags implements DomTagInterface {
 	/**
 	 * Construct the DOMtag.
-	 * @since 1.0.0
+	 * @since 1.2.0
 	 *
 	 * @access public
 	 * @param array|null $args (optional) -- The list of arguments.
 	 * @return string
 	 */
 	public static function tag(?array $args = null): string {
-		return parent::constructTag('fieldset', self::props(), $args);
+		return parent::constructTag('iframe', self::props(), $args);
 	}
 	
 	/**
 	 * The tag's props.
-	 * @since 1.0.0
+	 * @since 1.2.0
 	 *
 	 * @access public
 	 * @return array
@@ -31,7 +31,7 @@ class FieldsetTag extends \DomTags implements DomTagInterface {
 	public static function props(): array {
 		return array_merge(
 			parent::ALWAYS_WL,
-			array('name', 'disabled')
+			array('src', 'name', 'loading', 'allowfullscreen', 'width', 'height')
 		);
 	}
 }