+++ dom-tags/class-input-tag.php
@@ -10,6 +10,28 @@ namespace DomTags;
 
 class InputTag extends \DomTags implements DomTagInterface {
 	/**
+	 * Types of inputs.
+	 * @since 1.2.0
+	 *
+	 * @access private
+	 * @var array
+	 */
+	private const TYPES = array(
+		'button', 'checkbox', 'color', 'date', 'datetime-local', 'email', 'file',
+		'hidden', 'image', 'month', 'number', 'password', 'radio', 'range',
+		'reset', 'search', 'submit', 'tel', 'text', 'time', 'url', 'week'
+	);
+	
+	/**
+	 * The current type.
+	 * @since 1.2.0
+	 *
+	 * @access private
+	 * @var string
+	 */
+	private static $type;
+	
+	/**
 	 * Construct the DOMtag.
 	 * @since 1.0.0
 	 *
@@ -18,6 +40,18 @@ class InputTag extends \DomTags implements DomTagInterface {
 	 * @return string
 	 */
 	public static function tag(?array $args = null): string {
+		if(isset($args['type'])) {
+			if(!array_key_exists($args['type'], self::TYPES))
+				return 'Invalid input type!';
+		} else {
+			$args = array_merge(
+				array('type' => 'text'),
+				$args
+			);
+		}
+		
+		self::$type = $args['type'];
+		
 		return parent::constructTag('input', self::props(), $args);
 	}
 	
@@ -29,10 +63,80 @@ class InputTag extends \DomTags implements DomTagInterface {
 	 * @return array
 	 */
 	public static function props(): array {
-		return array_merge(
+		$props = array_merge(
 			array('type'),
 			parent::ALWAYS_WL,
-			array('name', 'maxlength', 'pattern', 'value', 'placeholder', 'checked', 'disabled', 'required', 'autofocus')
+			array('name', 'value', 'autocomplete', 'list')
+		);
+		
+		switch(self::$type) {
+			case 'checkbox':
+			case 'radio':
+				$props = array_merge(
+					$props,
+					array('checked', 'required')
+				);
+				break;
+			case 'date':
+				$props = array_merge(
+					$props,
+					array('min', 'max', 'step', 'pattern', 'required')
+				);
+				break;
+			case 'datetime-local':
+			case 'month':
+			case 'time':
+			case 'week':
+				$props = array_merge(
+					$props,
+					array('min', 'max', 'step', 'required')
+				);
+				break;
+			case 'email':
+				$props = array_merge(
+					$props,
+					array( 'placeholder', 'minlength', 'maxlength', 'pattern', 'multiple', 'required')
+				);
+				break;
+			case 'file':
+				$props = array_merge(
+					$props,
+					array('accept', 'multiple', 'required')
+				);
+				break;
+			case 'image':
+				$props = array_merge(
+					$props,
+					array('src', 'alt', 'width', 'height')
+				);
+				break;
+			case 'number':
+				$props = array_merge(
+					$props,
+					array('placeholder', 'min', 'max', 'step', 'required')
+				);
+				break;
+			case 'password':
+			case 'search':
+			case 'tel':
+			case 'text':
+			case 'url':
+				$props = array_merge(
+					$props,
+					array('placeholder', 'minlength', 'maxlength', 'pattern', 'required')
+				);
+				break;
+			case 'range':
+				$props = array_merge(
+					$props,
+					array('min', 'max', 'step')
+				);
+				break;
+		}
+		
+		return array_merge(
+			$props,
+			array('readonly', 'disabled', 'autofocus')
 		);
 	}
 }