+++ dom-tags/class-list-tag.php
@@ -19,6 +19,15 @@ class ListTag extends \DomTags implements DomTagInterface {
 	private const TAG_TYPES = array('ul', 'ol');
 	
 	/**
+	 * The current tag type.
+	 * @since 1.2.0
+	 *
+	 * @access private
+	 * @var string
+	 */
+	private static $tag_type;
+	
+	/**
 	 * Construct the DOMtag.
 	 * @since 1.0.0
 	 *
@@ -27,12 +36,12 @@ class ListTag extends \DomTags implements DomTagInterface {
 	 * @return string
 	 */
 	public static function tag(?array $args = null): string {
-		$type = self::TAG_TYPES[0];
+		self::$tag_type = self::TAG_TYPES[0];
 		
-		if(isset($args['type']) && in_array($args['type'], self::TAG_TYPES, true))
-			$type = $args['type'];
+		if(isset($args['tag_type']) && in_array($args['tag_type'], self::TAG_TYPES, true))
+			self::$tag_type = $args['tag_type'];
 		
-		return parent::constructTag($type, self::props(), $args);
+		return parent::constructTag(self::$tag_type, self::props(), $args);
 	}
 	
 	/**
@@ -43,6 +52,15 @@ class ListTag extends \DomTags implements DomTagInterface {
 	 * @return array
 	 */
 	public static function props(): array {
-		return parent::ALWAYS_WL;
+		$props = parent::ALWAYS_WL;
+		
+		if(self::$tag_type === 'ol') {
+			$props = array_merge(
+				$props,
+				array('type', 'start', 'reversed')
+			);
+		}
+		
+		return $props;
 	}
 }